workspaces/update_git_repo/requirements.txt
dagster
dagster-postgres
dagster-docker
python-dotenv
gitpython
docker